added verifcation of user_id to DELTE and PUT routes and fixed typos

+++ Controllers/totalBalance.js
@@ -14,27 +14,18 @@ const pool = new Pool({
 
 let client;
 let userAccountResults;
-
-//Example route for crud operations on users and their account details
-router.get("/join", async (req, res)=>{
-    try{
-        client = await pool.connect();
-        const fetchedUsers = await client.query("SELECT * FROM users FULL JOIN user_accounts ON users.id = user_accounts.user_id;");
-        res.status(200).json(fetchedUsers.rows);
-    }catch(error){
-        res.status(500).json({error:error.message});
-    }finally{
-        client.release();
-    }
-});
+let userIdResults;
 
 //Routes
+
+//GET all the existing transactions
 router.get("/", async (req, res)=>{
     try{
         client = await pool.connect(); //connect() establishes connections with Postgres
-        const fetchedUsers = await client.query("SELECT * FROM user_accounts;");
+        const fetchedUsers = await client.query("SELECT * FROM users FULL JOIN user_accounts ON users.id = user_accounts.user_id;");
         const formattedData = fetchedUsers.rows.map(row => ({
             id: row.id,
+            user_name: row.user_name,
             user_id: row.user_id,
             total_balance: row.total_balance,
             date_transaction: row.date_transaction.toISOString().split("T")[0], //The toISOString() method converst the Date object into a string, so we can use the split() method at the "T" and return the first section of that string array created by split() method which is the section with ONLY the date format
@@ -50,6 +41,7 @@ router.get("/", async (req, res)=>{
     }
 });
 
+//GET all the transactions of a SINGLE User
 router.get("/:id", async (req, res)=>{
     const {id} = req.params;
     try{
@@ -81,6 +73,7 @@ router.get("/:id", async (req, res)=>{
     }
 })
 
+//POST - Create transaction for SINGLE User
 router.post("/:id", async (req, res)=>{
     const {id} = req.params;
     const {income_transaction, expense_transaction} = req.body;
@@ -96,30 +89,31 @@ router.post("/:id", async (req, res)=>{
     }
 });
 
-router.put("/:transactionId", async (req, res)=>{
-    const {transactionId} = req.params;
+//PUT - Edit transaction (Frontend will restrict most RECENT transaction as only transaction available to edit)
+router.put("/:id", async (req, res)=>{
+    const {id} = req.params;
     const {user_id, income_transaction, expense_transaction} = req.body;
     try{
         client = await pool.connect();
 
-        const userIdResults = await client.query("SELECT * FROM user_accounts WHERE user_id = $1;", [user_id]);
+        userIdResults = await client.query("SELECT * FROM user_accounts WHERE user_id = $1;", [user_id]);
         if(userIdResults.rows.length === 0){
             res.status(403);
             throw new Error (`User id ${user_id} account details is a protected source`);
         }
 
-        userAccountResults = await client.query("SELECT * FROM user_accounts WHERE id = $1;", [transactionId]);
+        userAccountResults = await client.query("SELECT * FROM user_accounts WHERE id = $1;", [id]);
         if(userAccountResults.rows.length === 0){
             res.status(404);
-            throw new Error (`History of User transaction with id ${transactionId} does not exist, please provide valid transaction id`);
+            throw new Error (`History of User transaction with id ${id} does not exist, please provide valid transaction id`);
         }else{
-            await client.query("UPDATE user_accounts SET income_transaction = $1, expense_transaction = $2 WHERE id = $3;", [income_transaction, expense_transaction, transactionId]);
+            await client.query("UPDATE user_accounts SET income_transaction = $1, expense_transaction = $2 WHERE id = $3;", [income_transaction, expense_transaction, id]);
             res.status(200).json({message: "Record updated successfully"});
         }
     }catch(error){
-        if(res.statusCode === 404){
+        if(res.statusCode === 403){
             res.json({error:error.message});
-        }else if(res.statusCode === 403){
+        }else if(res.statusCode === 404){
             res.json({error:error.message});
         }else{
         res.status(500).json({error:error.message});
@@ -129,21 +123,32 @@ router.put("/:transactionId", async (req, res)=>{
     }
 });
 
+//DELETE - Delete transaction of SINGLE User (Frontend will require user to send user_id as confirmation for DELETION)
 router.delete("/:id", async (req, res)=>{
     const {id} = req.params;
+    const {user_id} = req.body;
     try{
         client = await pool.connect();
+
+        userIdResults = await client.query("SELECT * FROM user_accounts WHERE user_id = $1;", [user_id]);
+        if(userIdResults.rows.length === 0){
+            res.status(403);
+            throw new Error (`User id ${user_id} account details is a protected source`);
+        }
+
         userAccountResults = await client.query("SELECT * FROM user_accounts WHERE id = $1;", [id]);
         if(userAccountResults.rows.length === 0){
             //Condition is based on length of the array that the "rows" property has, which is a property from the object that is retrieved form the query (If lenght is 0, it means there are no records that result from the query)
             res.status(404);
-            throw new Error ("User account does not exist, please provide valid id");
+            throw new Error ("User transaction does not exist, please provide valid transaction id");
         }else{
             await client.query("DELETE FROM user_accounts WHERE id = $1;", [id]);
             res.status(200).json({message: `Record with id ${id} deleted successfully`});
         }  
     }catch(error){
-        if(res.statusCode === 404){
+        if(res.statusCode === 403){
+            res.json({error:error.message});
+        }else if(res.statusCode === 404){
             res.json({error:error.message});
         }else{
         res.status(500).json({error:error.message});
